+++ src/components/Radio/Radio.tsx
@@ -46,7 +46,7 @@ export interface RadioProps
   /** Offers the ability to show the radio's label onscreen or hide it. Refer
    * to the `labelText` property for more information. */
   showLabel?: boolean;
-  // /** Populates the value of the input. */
+  /** Populates the value of the input. */
   value?: string;
 }