+++ src/components/Button/buttonChangelogData.ts
@@ -20,6 +20,13 @@ export const changelogData: ChangelogData[] = [
     ],
   },
   {
+    date: "2025-05-22",
+    version: "3.6.2",
+    type: "Bug Fix",
+    affects: ["Styles"],
+    notes: ["Fixes the style transitions."],
+  },
+  {
     date: "2025-04-10",
     version: "3.6.0",
     type: "Update",