+++ src/components/MultiSelectGroup/multiSelectGroupChangelogData.ts
@@ -10,6 +10,22 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "Prerelease",
+    version: "Prerelease",
+    type: "Update",
+    affects: ["Styles"],
+    notes: [
+      "Removes use of `useNYPLBreakpoints` and replaces with equivalent CSS.",
+    ],
+  },
+  {
+    date: "2025-05-22",
+    version: "3.6.2",
+    type: "Update",
+    affects: ["Styles"],
+    notes: ["Uses CSS to decrease reliance on `useNYPLBreakpoints` hook."],
+  },
+  {
     date: "2024-07-03",
     version: "3.1.7",
     type: "Update",