+++ CHANGELOG.md
@@ -40,7 +40,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
   - `StatusBadge`'s `level` prop
   - `StyledList`'s "tag", "mini" values in `textSizesArray`
 - Removes `role="search"` from `Searchbar` wrapper.
-- Removes explicit `className` and `children` props in favor of expanded prop type definitions
+- Removes explicit `className` and `children` props in favor of expanded prop type definitions in all components.
 
 ## Prerelease